feat: optimize dependencies and features
