+++ manta-accounting/Cargo.toml
@@ -52,7 +52,7 @@ derivative = { version = "2.2.0", default-features = false, features = ["use_cor
 derive_more = { version = "0.99.16", default-features = false, features = ["add", "add_assign", "display", "from", "sum"] }
 indexmap = { version = "1.8.0", optional = true, default-features = false }
 manta-crypto = { path = "../manta-crypto", default-features = false }
-manta-util = { path = "../manta-util", features = ["alloc"] }
+manta-util = { path = "../manta-util", default-features = false, features = ["alloc"] }
 parking_lot = { version = "0.12.0", optional = true, default-features = false }
 rand = { version = "0.8.4", optional = true, default-features = false }
 rand_chacha = { version = "0.3.1", optional = true, default-features = false }
