+++ manta-pay/Cargo.toml
@@ -30,13 +30,12 @@ harness = false
 
 [[bin]]
 name = "generate_parameters"
-required-features = ["arkworks", "test"]
+required-features = ["groth16", "manta-util/std", "test"]
 
 [features]
 # Enable Arkworks Backend
 arkworks = [
     "ark-bls12-381",
-    "ark-crypto-primitives",
     "ark-ec",
     "ark-ed-on-bls12-381",
     "ark-ff",
@@ -47,7 +46,7 @@ arkworks = [
 ]
 
 # Enable Groth16 ZKP System
-groth16 = ["arkworks", "ark-groth16"]
+groth16 = ["ark-groth16", "ark-snark", "arkworks"]
 
 # TODO: Enable PLONK ZKP System
 # TODO: plonk = ["zk-garage-plonk"]
@@ -77,7 +76,6 @@ wallet = ["bip32", "manta-crypto/getrandom", "std"]
 [dependencies]
 aes-gcm = { version = "0.9.4", default-features = false, features = ["aes", "alloc"] }
 ark-bls12-381 = { version = "0.3.0", optional = true, default-features = false, features = ["curve"] }
-ark-crypto-primitives = { version = "0.3.0", optional = true, default-features = false, features = ["r1cs"] }
 ark-ec = { version = "0.3.0", optional = true, default-features = false }
 ark-ed-on-bls12-381 = { version = "0.3.0", optional = true, default-features = false, features = ["r1cs"] }
 ark-ff = { version = "0.3.0", optional = true, default-features = false }
@@ -85,12 +83,12 @@ ark-groth16 = { version = "0.3.0", optional = true, default-features = false }
 ark-r1cs-std = { version = "0.3.1", optional = true, default-features = false }
 ark-relations = { version = "0.3.0", optional = true, default-features = false }
 ark-serialize = { version = "0.3.0", optional = true, default-features = false }
+ark-snark = { version = "0.3.0", optional = true, default-features = false }
 ark-std = { version = "0.3.0", optional = true, default-features = false }
-bip32 = { version = "0.3.0", default-features = false, features = ["bip39", "secp256k1"], optional = true }
+bip32 = { version = "0.3.0", optional = true, default-features = false, features = ["bip39", "secp256k1"] }
 blake2 = { version = "0.10.0", default-features = false }
-derivative = { version = "2.2.0", default-features = false }
-generic-array = { version = "0.14.4", default-features = false }
-indexmap = { version = "1.8.0", default-features = false, optional = true }
+derivative = { version = "2.2.0", default-features = false, features = ["use_core"] }
+indexmap = { version = "1.8.0", optional = true, default-features = false }
 manta-accounting = { path = "../manta-accounting", default-features = false }
 manta-crypto = { path = "../manta-crypto", default-features = false }
 manta-util = { path = "../manta-util", default-features = false }
@@ -106,8 +104,5 @@ tungstenite = { version = "0.16.0", optional = true, default-features = false, f
 
 [dev-dependencies]
 criterion = "0.3.5"
-manta-accounting = { path = "../manta-accounting", features = ["test"] }
-manta-crypto = { path = "../manta-crypto", features = ["test"] }
-manta-pay = { path = ".", features = ["arkworks", "bip32", "groth16", "std", "test"] }
-rand = "0.8.4"
+manta-pay = { path = ".", features = ["wallet", "test"] }
 tempfile = "3.2.0"
