+++ manta-pay/src/signer/client/websocket.rs
